+++ packages/libp2p/src/connection-manager/index.ts
@@ -20,7 +20,7 @@ import type { Metrics } from '@libp2p/interface/metrics'
 import type { PeerId } from '@libp2p/interface/peer-id'
 import type { Peer, PeerStore } from '@libp2p/interface/peer-store'
 import type { Startable } from '@libp2p/interface/startable'
-import type { ConnectionManager } from '@libp2p/interface-internal/connection-manager'
+import type { ConnectionManager, OpenConnectionOptions } from '@libp2p/interface-internal/connection-manager'
 import type { TransportManager } from '@libp2p/interface-internal/transport-manager'
 
 const log = logger('libp2p:connection-manager')
@@ -65,6 +65,12 @@ export interface ConnectionManagerInit {
    * open connections above `minConnections`. (default: 100)
    */
   autoDialMaxQueueLength?: number
+
+  /**
+   * When we've failed to dial a peer, do not autodial them again within this
+   * number of ms. (default: 1 minute)
+   */
+  autoDialPeerRetryThreshold?: number
 
   /**
    * Sort the known addresses of a peer before trying to dial, By default public
@@ -152,10 +158,6 @@ export interface DefaultConnectionManagerComponents {
   transportManager: TransportManager
   connectionGater: ConnectionGater
   events: EventEmitter<Libp2pEvents>
-}
-
-export interface OpenConnectionOptions extends AbortOptions {
-  priority?: number
 }
 
 /**
@@ -492,7 +494,7 @@ export class DefaultConnectionManager implements ConnectionManager, Startable {
 
     const { peerId } = getPeerAddress(peerIdOrMultiaddr)
 
-    if (peerId != null) {
+    if (peerId != null && options.force !== true) {
       log('dial %p', peerId)
       const existingConnections = this.getConnections(peerId)