+++ packages/interface-internal/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+### [0.1.4](https://www.github.com/libp2p/js-libp2p/compare/interface-internal-v0.1.3...interface-internal-v0.1.4) (2023-08-16)
+
+
+### Features
+
+* **libp2p:** direct connection through relay protocol (DCUtR) ([#1928](https://www.github.com/libp2p/js-libp2p/issues/1928)) ([87dc7e9](https://www.github.com/libp2p/js-libp2p/commit/87dc7e9fc17becc4b5c3ce4f3febd28cf9f25c6e))
+
+### [0.1.3](https://www.github.com/libp2p/js-libp2p/compare/interface-internal-v0.1.2...interface-internal-v0.1.3) (2023-08-14)
+
+
+### Bug Fixes
+
+* update project config ([9c0353c](https://www.github.com/libp2p/js-libp2p/commit/9c0353cf5a1e13196ca0e7764f87e36478518f69))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^0.1.1 to ^0.1.2
+    * @libp2p/peer-collections bumped from ^4.0.2 to ^4.0.3
+
 ### [0.1.2](https://www.github.com/libp2p/js-libp2p/compare/interface-internal-v0.1.1...interface-internal-v0.1.2) (2023-08-05)