+++ packages/kad-dht/CHANGELOG.md
@@ -5,6 +5,50 @@
 
 * skip self-query if not running ([#479](https://github.com/libp2p/js-libp2p-kad-dht/issues/479)) ([7095290](https://github.com/libp2p/js-libp2p-kad-dht/commit/70952907a27fd8778773172059879656b4f08855))
 
+### [10.0.5](https://www.github.com/libp2p/js-libp2p/compare/kad-dht-v10.0.4...kad-dht-v10.0.5) (2023-08-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^4.0.4 to ^4.0.5
+
+### [10.0.4](https://www.github.com/libp2p/js-libp2p/compare/kad-dht-v10.0.3...kad-dht-v10.0.4) (2023-08-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^0.1.3 to ^0.1.4
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^4.0.3 to ^4.0.4
+
+### [10.0.3](https://www.github.com/libp2p/js-libp2p/compare/kad-dht-v10.0.2...kad-dht-v10.0.3) (2023-08-14)
+
+
+### Bug Fixes
+
+* **@libp2p/kad-dht:** update timeout ref type ([02b8932](https://www.github.com/libp2p/js-libp2p/commit/02b89323130f6d70a0f804f7f1a6adba81ea4d0a))
+* update project config ([9c0353c](https://www.github.com/libp2p/js-libp2p/commit/9c0353cf5a1e13196ca0e7764f87e36478518f69))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^2.0.2 to ^2.0.3
+    * @libp2p/interface bumped from ^0.1.1 to ^0.1.2
+    * @libp2p/interface-internal bumped from ^0.1.2 to ^0.1.3
+    * @libp2p/logger bumped from ^3.0.1 to ^3.0.2
+    * @libp2p/peer-collections bumped from ^4.0.2 to ^4.0.3
+    * @libp2p/peer-id bumped from ^3.0.1 to ^3.0.2
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^4.0.2 to ^4.0.3
+    * @libp2p/peer-id-factory bumped from ^3.0.2 to ^3.0.3
+    * @libp2p/peer-store bumped from ^9.0.2 to ^9.0.3
+
 ### [10.0.2](https://www.github.com/libp2p/js-libp2p/compare/kad-dht-v10.0.1...kad-dht-v10.0.2) (2023-08-05)