+++ packages/multistream-select/CHANGELOG.md
@@ -11,6 +11,21 @@
 
 * **dev:** bump aegir from 38.1.8 to 39.0.10 ([#70](https://github.com/libp2p/js-libp2p-multistream-select/issues/70)) ([f87b1c3](https://github.com/libp2p/js-libp2p-multistream-select/commit/f87b1c3505934ebeed6eff018af8d3042e7e6e06))
 
+### [4.0.2](https://www.github.com/libp2p/js-libp2p/compare/multistream-select-v4.0.1...multistream-select-v4.0.2) (2023-08-14)
+
+
+### Bug Fixes
+
+* update project config ([9c0353c](https://www.github.com/libp2p/js-libp2p/commit/9c0353cf5a1e13196ca0e7764f87e36478518f69))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^0.1.1 to ^0.1.2
+    * @libp2p/logger bumped from ^3.0.1 to ^3.0.2
+
 ### [4.0.1](https://www.github.com/libp2p/js-libp2p/compare/multistream-select-v4.0.0...multistream-select-v4.0.1) (2023-08-05)